backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pymupdf
openai
python-dotenv
httpx